AiHelper.spoon/requirements.txt
cohere
openai
pybars3